+++ mne/tests/test_epochs.py
@@ -6,6 +6,7 @@ import os.path as op
 from nose.tools import assert_true
 from numpy.testing import assert_array_equal, assert_array_almost_equal
 import numpy as np
+import copy as cp
 
 from mne import fiff, Epochs, read_events, pick_events
 from mne.epochs import bootstrap
@@ -216,6 +217,33 @@ def test_crop():
     assert_array_equal(data2, data_normal[:, :, tmask])
     assert_array_equal(data3, data_normal[:, :, tmask])
 
+def test_resample():
+    """Test of resample of epochs
+    """
+    epochs = Epochs(raw, events[:5], event_id, tmin, tmax, picks=picks,
+                    baseline=(None, 0), preload=True,
+                    reject=reject, flat=flat)
+    data_normal = cp.deepcopy(epochs.get_data())
+    times_normal = cp.deepcopy(epochs.times)
+    sfreq_normal = epochs.info['sfreq']
+    # upsample by 2
+    epochs.resample(sfreq_normal*2)
+    data_up = cp.deepcopy(epochs.get_data())
+    times_up = cp.deepcopy(epochs.times)
+    sfreq_up = epochs.info['sfreq']
+    # downsamply by 2, which should match
+    epochs.resample(sfreq_normal)
+    data_new = cp.deepcopy(epochs.get_data())
+    times_new = cp.deepcopy(epochs.times)
+    sfreq_new = epochs.info['sfreq']
+
+    assert_true(data_up.shape[2] == 2*data_normal.shape[2])
+    assert_true(sfreq_up == 2*sfreq_normal)
+    assert_true(sfreq_new == sfreq_normal)
+    assert_true(len(times_up) == 2*len(times_normal))
+    assert_array_almost_equal(times_new, times_normal, 10)
+    assert_true(data_up.shape[2] == 2*data_normal.shape[2])
+    assert_array_almost_equal(data_new, data_normal, 2)
 
 def test_bootstrap():
     """Test of bootstrapping of epochs